Refactored - Created new LocalePreference class for storing/loading user locale state.
