+++ app/src/org/commcare/utils/ChangeLocaleUtil.java
@@ -1,16 +1,8 @@
 package org.commcare.utils;
 
-import android.content.Context;
-import android.content.res.Configuration;
-import android.content.res.Resources;
-import android.os.Build;
-
-import org.commcare.dalvik.BuildConfig;
 import org.javarosa.core.services.locale.Localization;
 import org.javarosa.core.services.locale.Localizer;
 import org.javarosa.core.util.NoLocalizedTextException;
-
-import java.util.Locale;
 
 public class ChangeLocaleUtil {
 
@@ -48,41 +40,4 @@ public class ChangeLocaleUtil {
         String[] rawDefaultRemoved = getLocaleCodes();
         return translateLocales(rawDefaultRemoved);
     }
-
-    /**
-     * Creates new context with updated layout configuration based on locale from {@link Localization#getCurrentLocale}.
-     *
-     * @param context old context.
-     * @return updated context.
-     */
-    public static Context setLocale(Context context) {
-        try {
-            return updateResources(context, Localization.getCurrentLocale());
-        } catch (Exception ignored) {
-        }
-        return context;
-    }
-
-    private static Context updateResources(Context context, String language) {
-        Locale locale = new Locale(language);
-        Locale.setDefault(locale);
-
-        Resources res = context.getResources();
-        Configuration config = new Configuration(res.getConfiguration());
-        if (Build.VERSION.SDK_INT >= 17) {
-            config.setLocale(locale);
-            context = context.createConfigurationContext(config);
-//            res.updateConfiguration(config, res.getDisplayMetrics());
-        } else {
-            config.locale = locale;
-            res.updateConfiguration(config, res.getDisplayMetrics());
-        }
-        return context;
-    }
-
-    public static Locale getLocale(Resources res) {
-        Configuration config = res.getConfiguration();
-        return Build.VERSION.SDK_INT >= 24 ? config.getLocales().get(0) : config.locale;
-    }
-
 }
