+++ app/src/org/commcare/views/dialogs/PaneledChoiceDialog.java
@@ -13,8 +13,7 @@ import android.widget.ListView;
 import android.widget.TextView;
 
 import org.commcare.dalvik.R;
-import org.commcare.preferences.MainConfigurablePreferences;
-import org.commcare.utils.LayoutDirectionUtilCompat;
+import org.commcare.preferences.LocalePreferences;
 
 /**
  * An implementation of CommCareAlertDialog for use in any instance in which the user is being
@@ -73,7 +72,7 @@ public class PaneledChoiceDialog extends CommCareAlertDialog {
         if (item.iconResId != -1) {
             Drawable icon = ContextCompat.getDrawable(context, item.iconResId);
             if (iconToLeft) {
-                if (MainConfigurablePreferences.isLocaleRTL())
+                if (LocalePreferences.isLocaleRTL())
                     choicePanel.setCompoundDrawablesWithIntrinsicBounds(null, null, icon, null);
                 else
                     choicePanel.setCompoundDrawablesWithIntrinsicBounds(icon, null, null, null);
